refactor(SCRUM-166): image, ai 패키지 global로 이동
